keeping created_at trait

+++ server/actions/update-user.js
@@ -29,18 +29,19 @@ export default function updateUser({ client, service, ship }: Object, messages: 
   };
 
   const { syncAgent } = service;
-  return Promise.all(messages.map(({ user, segments, events }) => {
+  return Promise.all(messages.map(({ user, segments, events, changes }) => {
     const userPromises = [];
-
-    if (_.intersection(segments.map(s => s.id), filterSegments).length > 0) {
-      userPromises.push(syncAgent.sendAllUserProperties(user, segments));
-    } else if (userDeletionEnabled && !_.get(user, "traits_customerio/deleted_at")) {
-      userPromises.push(syncAgent.deleteUser(user));
-    } else {
-      client.asUser(user).logger.info("outgoing.user.skip", {
-        id: syncAgent.getUsersCustomerioId(user),
-        reason: "User is not included in synchronized segments setting and: user deletion is disabled or he was deleted"
-      });
+    if (!_.get(changes, "user['traits_customerio/created_at'][1]", false) || !_.get(changes, "user['traits_customerio/deleted_at'][1]", false)) {
+      if (_.intersection(segments.map(s => s.id), filterSegments).length > 0) {
+        userPromises.push(syncAgent.sendAllUserProperties(user, segments));
+      } else if (userDeletionEnabled && !_.get(user, "traits_customerio/deleted_at")) {
+        userPromises.push(syncAgent.deleteUser(user));
+      } else {
+        client.asUser(user).logger.info("outgoing.user.skip", {
+          id: syncAgent.getUsersCustomerioId(user),
+          reason: "User is not included in synchronized segments setting and: user deletion is disabled or he was deleted"
+        });
+      }
     }
 
     return Promise.all(userPromises).then(() =>
